+++ api/src/controller/categoriesController.js
@@ -1,0 +1,24 @@
+const { Categories } = require("../db.js");
+
+async function allCategories(req, res) {
+  try {
+    let categoriesInDb = await Categories.findAll();
+
+    if (categoriesInDb.length) return res.status(201).json(categoriesInDb);
+  } catch (error) {
+    res.status(404).json(error);
+  }
+}
+
+async function createCategories(req, res) {
+  let { name } = req.body;
+
+  const newCategory = await Categories.create({ name });
+
+  res.status(201).json(newCategory);
+}
+
+module.exports = {
+  allCategories,
+  createCategories,
+};
