+++ api/src/routes/index.js
@@ -1,17 +1,15 @@
 const { Router } = require("express");
-const categoriesRoutes = require("./categoriesRoutes");
-const productsRoutes = require("./productsRoutes");
-const productsById = require("./productsById");
-const registerRoutes = require("./registerRoutes");
-const loginRoutes = require("./loginRoutes");
-const userRoutes = require("./usersRoutes");
-
+const products = require("../routes/productsRoutes");
+const categories = require("../routes/categoriesRoutes");
+const register = require("../routes/registerRoutes");
+const login = require("../routes/loginRoutes");
+const users = require("../routes/usersRoutes");
 const router = Router();
-router.use("/register", registerRoutes);
-router.use("/login", loginRoutes);
-router.use("/users", userRoutes);
-router.use("/products", productsRoutes);
-router.use("/product", productsById);
-router.use("/categories", categoriesRoutes);
+
+router.use("/products", products);
+router.use("/categories", categories);
+router.use("/register", register);
+router.use("/login", login);
+router.use("/users", users);
 
 module.exports = router;
